requirements.txt
aiogram>=3.0.0b9
python-dotenv
pytest
